+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # nest-indexer 체인지로그
 
+## [develop] - [v1.0.2] (2026-05-13 01:23)
+### Added
+- Entities 추가(block, transaction, log, indexer-checkpoint)
+
 ## [develop] - [v1.0.1] (2026-05-11)
 ### Added
 - 클린 아키텍처 구조 기본 디렉토리 추가(src/application, controllers, entities, frameworks)